+++ trax/layers/__init__.py
@@ -122,4 +122,5 @@ AveragePooling = layer_configure(AveragePooling)
 LinearPooling = layer_configure(LinearPooling)
 LinearUpsampling = layer_configure(LinearUpsampling)
 NoUpsampling = layer_configure(NoUpsampling)
+NaiveUpsampling = layer_configure(NaiveUpsampling)
 AttentionResampling = layer_configure(AttentionResampling)
