+++ trax/layers/research/resampling.py
@@ -61,6 +61,10 @@ def LinearUpsampling(shorten_factor, d_model, *args, dropout=0.0, mode='train',
   )
 
 
+def NaiveUpsampling(shorten_factor, d_model, *args, **kwargs):
+  return core.Fn('Repeat', lambda x: jnp.repeat(x, shorten_factor, axis=1))
+
+
 def NoUpsampling(shorten_factor, d_model, *args, **kwargs):
   del d_model, args, kwargs
 
@@ -68,12 +72,12 @@ def NoUpsampling(shorten_factor, d_model, *args, **kwargs):
       (x.shape[0], x.shape[1] * shorten_factor, x.shape[2]), dtype=x.dtype))
 
 
-def _FeedForwardBlock(d_model,
-                      d_ff,
-                      dropout,
-                      dropout_shared_axes,
-                      mode,
-                      activation):
+def FeedForwardBlock(d_model,
+                     d_ff,
+                     dropout,
+                     dropout_shared_axes,
+                     mode,
+                     activation):
   # We copy the ff block function because we cannot import it from models
   return [
       core.Dense(d_ff),
@@ -95,7 +99,7 @@ def AttentionResampling(shorten_factor, d_model, is_upsampling, d_ff, n_heads,
       total_pooling, n_heads=n_heads, dropout=dropout,
       mode=mode)
 
-  feed_forward = _FeedForwardBlock(
+  feed_forward = FeedForwardBlock(
       d_model, d_ff, dropout, dropout_shared_axes, mode, ff_activation)
 
   resampling = resampling_fn(shorten_factor, d_model,
